+++ SparseSDF/SDFVoxCb.h
@@ -1,8 +1,6 @@
 #ifndef SDF_VOX_CB
 #define SDF_VOX_CB
 
-#include <unordered_map>
-
 #include "AdapDF.h"
 #include "TrigMesh.h"
 #include "VoxCallback.h"
@@ -13,7 +11,7 @@
 // also build mapping from sparse node to data index.
 struct TrigListVoxCb : public VoxCallback {
   virtual void operator()(unsigned x, unsigned y, unsigned z,
-                          size_t trigIdx) override;
+                          unsigned long long trigIdx) override;
   TrigMesh* m = nullptr;
   AdapDF* df = nullptr;
